Fix a grammatical error
The corrected sentence is being quoted in a DESC pedagogical note and there's dilemma about whether to quote it as it is with a [sic] included to denote the grammatical error or to silently fix the grammar. It would be cleaner to fix the error at the source so that the quote does not need to worry about it.